# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "医疗知识学习平台 API",
    "version": "1.0.0",
    "description": "医疗知识学习平台 API 文档"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "GenericSuccessMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "GenericErrorMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "错误信息描述"
          },
          "error": {
            "type": "string",
            "nullable": true,
            "description": "详细错误内容 (可选)"
          }
        },
        "example": {
          "message": "请求处理失败。",
          "error": "具体错误原因... (可选)"
        }
      },
...
      "Tag": {
        "type": "object",
        "properties": {
          "tagID": {
            "type": "string",
            "format": "uuid",
            "description": "标签ID"
          },
          "tagName": {
            "type": "string",
            "description": "标签名称",
            "example": "React"
          },
          "articleCount": {
            "type": "integer",
            "description": "使用该标签的文章数量",
            "example": 15
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          }
        }
      },
      "BriefArticleInfo": {
        "type": "object",
        "properties": {
          "articleID": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string",
            "nullable": true
          },
          "coverImageURL": {
            "type": "string",
            "format": "url",
            "nullable": true
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "TagWithArticles": {
        "type": "object",
        "properties": {
          "tagID": {
            "type": "string",
            "format": "uuid"
          },
          "tagName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "articles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BriefArticleInfo"
            },
            "description": "使用此标签的已发布文章列表"
          }
        }
      },
      "TagCreatePayload": {
        "type": "object",
        "required": [
          "tagName"
        ],
        "properties": {
          "tagName": {
            "type": "string",
            "description": "新标签的名称",
            "example": "JavaScript ES2023"
          }
        }
      },
      "TagUpdatePayload": {
        "type": "object",
        "required": [
          "tagName"
        ],
        "properties": {
          "tagName": {
            "type": "string",
            "description": "标签的新名称",
            "example": "ECMAScript 2023"
          }
        }
      },
      "TagResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "tag": {
            "$ref": "#/components/schemas/Tag"
          }
        }
      },
      "DeleteTagErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "标签正在被使用，无法删除!"
          },
          "usageCount": {
            "type": "integer",
            "example": 5
          }
        }
      },
      "PopularTagStat": {
        "type": "object",
        "properties": {
          "tagID": {
            "type": "string",
            "format": "uuid"
          },
          "count": {
            "type": "integer",
            "description": "使用次数"
          },
          "tag": {
            "type": "object",
            "properties": {
              "tagName": {
                "type": "string",
                "description": "标签名称"
              }
            }
          }
        }
      },
      "UnusedTagInfo": {
        "type": "object",
        "properties": {
          "tagID": {
            "type": "string",
            "format": "uuid"
          },
          "tagName": {
            "type": "string"
          }
        }
      },
      "TagStatsResponse": {
        "type": "object",
        "properties": {
          "popularTags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PopularTagStat"
            }
          },
          "totalTags": {
            "type": "integer",
            "example": 100
          },
          "unusedTagsCount": {
            "type": "integer",
            "example": 10
          },
          "unusedTags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UnusedTagInfo"
            }
          }
        }
      },
...
      "KnowledgeArticle": {
        "type": "object",
        "properties": {
          "articleID": {
            "type": "string",
            "format": "uuid",
            "description": "知识文章ID"
          },
          "title": {
            "type": "string",
            "description": "标题"
          },
          "summary": {
            "type": "string",
            "nullable": true,
            "description": "简介"
          },
          "coverImageURL": {
            "type": "string",
            "format": "url",
            "nullable": true,
            "description": "封面图片链接"
          },
          "richTextContent": {
            "type": "string",
            "nullable": true,
            "description": "富文本内容"
          },
          "videoURL": {
            "type": "string",
            "format": "url",
            "nullable": true,
            "description": "视频链接"
          },
          "status": {
            "type": "string",
            "enum": [
              "Draft",
              "PendingReview",
              "Published",
              "Archived",
              "Rejected"
            ],
            "description": "状态"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "发布时间"
          },
          "version": {
            "type": "integer",
            "description": "当前版本号"
          },
          "averageRating": {
            "type": "number",
            "format": "float",
            "nullable": true,
            "description": "平均评分"
          },
          "viewCount": {
            "type": "integer",
            "description": "浏览次数"
          },
          "categoryID": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "所属分类ID"
          },
          "authorStaffID": {
            "type": "string",
            "format": "uuid",
            "description": "作者Staff ID"
          },
          "parentArticleID": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "父文章ID (用于系列文章)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "更新时间"
          }
        }
      },
      "ArticleAuthor": {
        "type": "object",
        "properties": {
          "staffID": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string"
          },
          "avatarURL": {
            "type": "string",
            "format": "url",
            "nullable": true
          }
        }
      },
      "ArticleCategorySummary": {
        "type": "object",
        "properties": {
          "categoryID": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "KnowledgeArticleBrief": {
        "allOf": [
          {
            "$ref": "#/components/schemas/KnowledgeArticle"
          }
        ],
        "type": "object",
        "properties": {
          "author": {
            "$ref": "#/components/schemas/ArticleAuthor"
          },
          "category": {
            "$ref": "#/components/schemas/ArticleCategorySummary",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tag"
            }
          }
        }
      },
      "KnowledgeArticleDetailed": {
        "allOf": [
          {
            "$ref": "#/components/schemas/KnowledgeArticleBrief"
          }
        ],
        "type": "object",
        "properties": {
          "isBookmarked": {
            "type": "boolean",
            "description": "当前用户是否已收藏此文章"
          },
          "relatedArticles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BriefArticleInfo"
            },
            "nullable": true
          },
          "parentArticle": {
            "type": "object",
            "nullable": true,
            "properties": {
              "articleID": {
                "type": "string",
                "format": "uuid"
              },
              "title": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/admin/tags/{tagId}": {
      "delete": {
        "summary": "删除标签 (需要管理员权限)",
        "tags": [
          "Admin - TagManagement"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "tagId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "要删除的标签ID"
          }
        ],
        "responses": {
          "200": {
            "description": "标签删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessMessage"
                }
              }
            }
          },
          "400": {
            "description": "标签正在被使用，无法删除",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteTagErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "无权限操作 (非管理员)"
          },
          "404": {
            "description": "标签不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
...
    "/api/v1/articles/{articleId}/tags": {
      "post": {
        "summary": "为文章添加标签 (工作人员)",
        "tags": [
          "Client - KnowledgeArticles"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "articleId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "文章ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "tagIds"
                ],
                "properties": {
                  "tagIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "description": "标签ID数组"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "标签添加成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "article": {
                      "$ref": "#/components/schemas/KnowledgeArticleDetailed"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "标签ID列表不能为空或标签不存在"
          },
          "403": {
            "description": "无权限 (非工作人员)"
          },
          "404": {
            "description": "文章不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/api/v1/articles/{articleId}/tags/{tagId}": {
      "delete": {
        "summary": "从文章移除标签 (工作人员)",
        "tags": [
          "Client - KnowledgeArticles"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "articleId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "文章ID"
          },
          {
            "in": "path",
            "name": "tagId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "标签ID"
          }
        ],
        "responses": {
          "200": {
            "description": "标签移除成功"
          },
          "403": {
            "description": "无权限 (非工作人员)"
          },
          "404": {
            "description": "文章或标签不存在，或文章没有此标签"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
...
    "/api/v1/tags/": {
      "get": {
        "summary": "获取所有标签列表",
        "tags": [
          "Client - Tags"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "按标签名称搜索 (可选)"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "enum": [
                "tagName",
                "createdAt",
                "articleCount"
              ],
              "default": "tagName"
            },
            "description": "排序字段 (tagName, createdAt, articleCount - articleCount may not be directly sortable by DB without specific query adjustments)"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ],
              "default": "ASC"
            },
            "description": "排序顺序"
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取标签列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tag"
                  }
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericErrorMessage"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "创建新标签 (需要工作人员权限)",
        "tags": [
          "Client - Tags"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagCreatePayload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "标签创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误或标签已存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericErrorMessage"
                }
              }
            }
          },
          "403": {
            "description": "无权限操作 (非工作人员)"
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericErrorMessage"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tags/{tagId}": {
      "get": {
        "summary": "获取特定标签的详情及其关联的文章",
        "tags": [
          "Client - Tags"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "tagId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "标签的ID"
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取标签详情",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagWithArticles"
                }
              }
            }
          },
          "404": {
            "description": "标签不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericErrorMessage"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericErrorMessage"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "更新标签名称 (需要工作人员权限)",
        "tags": [
          "Client - Tags"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "tagId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "要更新的标签ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagUpdatePayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "标签更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误或新标签名称已存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericErrorMessage"
                }
              }
            }
          },
          "403": {
            "description": "无权限操作 (非工作人员)"
          },
          "404": {
            "description": "标签不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericErrorMessage"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericErrorMessage"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tags/stats/usage": {
      "get": {
        "summary": "获取标签使用统计 (需要工作人员权限)",
        "tags": [
          "Client - Tags"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取标签统计数据",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagStatsResponse"
                }
              }
            }
          },
          "403": {
            "description": "无权限操作 (非工作人员)"
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericErrorMessage"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}